refactor(#101): extract downloadReport() to lib/reports.ts
Both fetch-reports and get-report-data had their own downloadReport()
function with overlapping HTTPS-download + CSV-parse logic. The two
copies diverged over time — fetch-reports grew retry/backoff and
parseCsvAndExtractRange-based minDate/maxDate extraction, while
get-report-data stayed single-shot and parsed CSV separately inline.

Moves three primitives to lib/reports.ts:
- safeTmpReportPath(reportId)  — /tmp filename sanitizer (was in fetch-reports)
- downloadOnce(url, token, dest) — single HTTPS GET, no retry (was in fetch-reports)
- downloadReport(report, auth, opts?) — retry-wrapped download + CSV parse
  - Accepts an optional tmpPath override so callers that own their temp
    directory (get-report-data) can keep their existing path scheme.
  - Returns {csvData, headers, data, minDate, maxDate} unconditionally —
    get-report-data previously parsed CSV separately to extract minDate/
    maxDate; now it consumes the helper's return directly.

Behavior change for get-report-data: gains retry/backoff on HTTP 429
and transient network errors (was: single attempt). Strict improvement.

Also drops unused imports in both commands (https, createWriteStream,
unlinkSync, parseCsvAndExtractRange, path).
